Data.CallStack

Documentation

type HasCallStack = ?callStack :: CallStack

Request a CallStack.

NOTE: The implicit parameter ?callStack :: CallStack is an implementation detail and should not be considered part of the CallStack API, we may decide to change the implementation in the future.

Since: base-4.9.0.0

type CallStack = [(String, SrcLoc)]

data SrcLoc

A single location in the source code.

Since: base-4.8.1.0

Constructors

SrcLoc

Fields

srcLocPackage :: [Char]

srcLocModule :: [Char]

srcLocFile :: [Char]

srcLocStartLine :: Int

srcLocStartCol :: Int

srcLocEndLine :: Int

srcLocEndCol :: Int

callStack :: HasCallStack => CallStack

callSite :: HasCallStack => Maybe (String, SrcLoc)